Move ActivationCapsule to run_context.utils
