+++ neuro_san/internals/run_context/utils/activation_capsule.py
